Export function for updating auth docs
The process of writing documents to the authentication DB varies
depending whether the authentication DB is on the admin or the
clustered interface. Authentication handlers should therefore
abstract the details away and provide a generic update_doc
function.

This commit adds an update_auth_doc function to the chttpd auth
cache which proxies to fabric:update_doc/3.

COUCHDB-2452 2/5