Dynamic endpoints handlers
chttpd hardcoded handlers are replaced with with dynamic url handlers,
which are functions in a dynamically created module, assembled from
priv/chttpd_handler.cfg files from all interested applications.

For couched these are currently: chttpd, mem3 and global_changes.

This is a special branch that pull the prepared branches (of the same name)
of chttpd, mem3 and global_changes, which have the config files.

Run 'make check' to test the handlers. They are implicitly tested by other
tests and there is a whitebox callback test chttpd_handler_callback_test,
which calls the dynamic functions directly and mocks the funs that they
return. This tests precisely the relationship between the *_handler function
clauses and the returned function.

There are more tests to come, both testing the endpoints via http, and the
configuration assembly.

BugzID: 27037