[squash] simplify IsUsersDb check
I realised the second part of the orelse is not needed here.
Because this code is in chttpd we will only be dealing with the
cluster-level DB name, not internal shard names, so there is
never a need to call mem3:dbname. We still need to convert from
binary to list though.

The replication DB check is also removed as that information is
not used here.

To squash into 9b5406b4b6d39ec5df6d1061c270f7a90e797a67